% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readfoundry-package.R
\docType{package}
\name{readfoundry-package}
\alias{readfoundry}
\alias{readfoundry-package}
\title{readfoundry: Wrapper for accesing data from Foundry SparkSQL server.}
\description{
Package consists of a single class that enables requests to be sent to the Foundry SparkSQL server. Requests can be sent ever as raw SparkSQL or via a `dbplyr` like interface. User authentication is handled by a bearer token generated by Foundry.
}
\author{
\strong{Maintainer}: Matthew Katz \email{matthew.katz@dhsc.gov.uk}

}
\keyword{internal}
